Be able to add multiple methods to pdf to text extractor
