+++ src/extractors/text_to_text_extractor/methods/RegexSubtractionMethod.py
@@ -2,12 +2,13 @@ import re
 
 from data.ExtractionData import ExtractionData
 from data.PredictionSample import PredictionSample
-from extractors.text_to_text_extractor.TextToTextMethod import TextToTextMethod
 
 from tdda import *
 
+from extractors.ToTextExtractorMethod import ToTextExtractorMethod
 
-class RegexSubtractionMethod(TextToTextMethod):
+
+class RegexSubtractionMethod(ToTextExtractorMethod):
 
     def train(self, extraction_data: ExtractionData):
         front_subtraction = [
