+++ src/scripts/check_performance.py
@@ -5,7 +5,7 @@ import pandas as pd
 from datetime import datetime
 
 from performance.Results import Results
-from extractors.text_to_text_extractor.TextToTextMethod import TextToTextMethod
+from extractors.text_to_text_extractor.TextToTextMethod import BaseMethod
 from extractors.text_to_text_extractor.methods.MT5TrueCaseEnglishSpanishMethod import MT5TrueCaseEnglishSpanishMethod
 
 
@@ -14,7 +14,7 @@ class CheckPerformance:
 
     TENANT = "check_performance"
 
-    METHODS: list[Type[TextToTextMethod]] = [
+    METHODS: list[Type[BaseMethod]] = [
         # T5ZeroShot,
         MT5TrueCaseEnglishSpanishMethod,
         # MT5EnglishSpanishMethod,
@@ -71,7 +71,7 @@ class CheckPerformance:
             accuracies = list()
             method_names = list()
 
-            train, test = TextToTextMethod.get_train_test(self.semantic_information_data, self.training_length)
+            train, test = BaseMethod.get_train_test(self.semantic_information_data, self.training_length)
 
             for self.current_method in self.METHODS:
                 self.run_method(accuracies)
@@ -110,7 +110,7 @@ class CheckPerformance:
         self.best_results.write_results()
 
     def write_mistakes(self):
-        _, performance_test_set = TextToTextMethod.get_train_test(self.semantic_information_data, self.training_length)
+        _, performance_test_set = BaseMethod.get_train_test(self.semantic_information_data, self.training_length)
         correct_path = "../performance_results/mistakes/"
         correct_path += f"{self.training_length}_{self.current_method_name}_{self.current_dataset}_correct.txt"
         mistakes_path = "../performance_results/mistakes/"
